+++ ThePhasesOfCraftship/2_best_practice_first/strategicDesignPart1/exercises/1_RefactoringTo4Tiers/end/src/dtos/students.ts
@@ -1,5 +1,5 @@
-import { isMissingKeys, isUUID } from "../utils/api_utils";
-import { InvalidRequestBodyError } from "../utils/exceptions";
+import { isMissingKeys, isUUID } from "../shared/utils";
+import { InvalidRequestBodyError } from "../shared/exceptions";
 
 class CreateStudentDTO {
     constructor(public name: string) {}
