+++ ThePhasesOfCraftship/2_best_practice_first/strategicDesignPart1/exercises/1_RefactoringTo4Tiers/end/src/index.ts
@@ -1,6 +1,6 @@
 import express from 'express';
 import { AssignmentsRouter, ClassesRouter, StudentsRouter } from './controllers';
-import { enableGracefulShutdown } from './utils/server';
+import { enableGracefulShutdown } from './shared/server';
 
 
 const app = express();
